[dev] reduce actor id size

+++ system/actor/ActorId.h
@@ -11,24 +11,24 @@ namespace cube {
     struct ActorId {
         using NotFound = ActorId;
 
-        uint32_t _id;
-        uint32_t _index;
+        uint16_t _id;
+        uint16_t _index;
     public:
         ActorId() : _id(0), _index(0) {}
-        ActorId(uint32_t const id, uint32_t const index)
+        ActorId(uint16_t const id, uint16_t const index)
                 : _id(id), _index(index) {}
         ActorId(ActorId const &) = default;
 
-        ActorId(uint64_t const id) {
-            *reinterpret_cast<uint64_t *>(this) = id;
+        ActorId(uint32_t const id) {
+            *reinterpret_cast<uint32_t *>(this) = id;
         }
 
-        inline operator const uint64_t &() const {
-            return *reinterpret_cast<uint64_t const *>(this);
+        inline operator const uint32_t &() const {
+            return *reinterpret_cast<uint32_t const *>(this);
         }
 
         inline bool operator!=(ActorId const &rhs) const {
-            return static_cast<uint64_t>(*this) != static_cast<uint64_t>(rhs);
+            return static_cast<uint32_t>(*this) != static_cast<uint64_t>(rhs);
         }
     };
